Quote.java
Write a program Quote.java which reads a line of text (using class
Scanner ) and then prints the same line as a quote (that is inside " ").
An example of an execution:
Write a line of text:  I wish I was a punk rocker with flowers in my
hair.
Quote: "I wish I was a punk rocker with flowers in my hair."