BMI.java
Write a program BMI.java which computes the BMI (Body Mass Index) for a
person. The program will read length and weight from the keyboard and
then present the result as output. The BMI is computed as
weight/(length)^2, where the length is given in meters and the weight in
kilograms. An example of an execution:
Give your length in meters: 1,83
Give your weight in kilograms: 83
Your BMI is: 25