Time.java
Write a program Time.java, which reads a number of seconds (an integer)
and then prints the same amount of time given in hours, minutes and
seconds. An example of an execution:

Give a number of seconds: 9999
This corresponds to: 2 hours, 46 minutes and 39 seconds.

Hint: Use integer division and the modulus (remainder) operator.