ShortName.java
Write a program ShortName.java, reading a first name and a last name
(given name and family name) as two Strings. The output should consist
of the first letter of the first name followed by a dot and a space,
followed by the first four letters of the last name. An example of an
execution:

First name: Anakin
Last name: Skywalker
Short name: A. Skyw

Hint: Use methods of the String class.
What will happen if the last name consists of less than four letters?