Distance.java
Write a program Distance.java which reads two coordinates in the form
(x,y) and then computes the distance between the points, using the
formula

distance = Sqrt( (x1-x2)^2 + (y1-y2)^2 )

Sqrt() means "the square root of" and ^ means "raised to". The answer
should be presented with three decimal digits.
Hint: Use the class java.lang.Math for the computations.